Add the main files and directories
Add the index.html file, the style directory with style.css and
reset.css files and the README.md files. Fill index.html head
element and reset.css file.

With the exception of README.md, these are all the files needed
to start working on the project, making this a good point for a
first commit.