Add image file and HTML,CSS code
Create the image directory and add a picture of SGA called SGA.jpeg.
Fill the body element of the index.html and the CSS code to style.css.

The picture is a required element for the project, hence its use.
The HTML and CSS code fills the remaining requirements for the project.