Add more pictures, a font and HTML, CSS code
Add pictures of Grothendieck from his childhood, youth, 1951, 1970,
Vietnam, 1988 and a picture for the header. Add a Tex font for the
page. Improve the content and add more HTML, CSS for semantics and
visual.

The pictures will help portray the history of Grothendieck. The font
is just a personal customization to let the page be more mathemati
cally designed. The improved content is probably final, but will add
some final touchs later if needed. Finally, the addition of the HTML,
CSS code are to improve the semantics and visial of the page (still
needs to add more).