Finish HTML, CSS coding and do some refactoring
Add the footer, some content, some elements ids and finish editing
the HTML code. Add the root styles, footer styles, media queries
and finish editing the CSS code. Finally do some refactoring of
the HTML and CSS.

With the footer I take credit for the page and present the links to
my profile in other networks exposing my work. The content addition
helps improve the overall flow of the page and the ids addtions are
used in the CSS. The root styles are used to define the navbar
height variables and set the base font-size for the others elements.
Footer styles are used to create a better footer layout, matching
the navbar and header layouts. The media queries are used to
improve the site responsiveness having two basic layouts, one for
portraits and one for landscapes. Finally, the refactoring cleans
some of the code, mainly the CSS code redundancy.