Change images, font links &  add README.md content
Change images src from a file in the repository to a link on GitHub,
change the CM Serif font href to a link on JSDelivr. Also add a
little description of the repository content on the README.md.

The change on the images src and the font href are made so that
the codepen of the project works properly. The content on README.md
will help people understand the objectives of the project.